requirements.txt
asgiref==3.8.1
Django==4.2.11
django-cors-headers==4.3.1
djangorestframework==3.15.1
mysqlclient==2.2.4
pillow==10.3.0
pytz==2024.1
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
